Minor bugfixes.
